examples/generators/starter_python_smart_contract_typescript/pyproject.toml
[tool.poetry]
name = "starter_python_smart_contract_typescript"
version = "0.1.0"
description = "Algorand smart contracts"
authors = ["None <None>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
algokit-utils = "^2.3.0"
python-dotenv = "^1.0.0"
algorand-python = "^2.0.0"
algorand-python-testing = "^0.3.0"

[tool.poetry.group.dev.dependencies]
puyapy = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

